Переменовал везде Event (events) в Point (points) и вынес константы
