+++ src/const.js
@@ -1,0 +1,12 @@
+const POINT_COUNT = 3;
+
+const PlaceTemplate = {
+  BEFOREBEGIN: `beforebegin`,
+  AFTERBEGIN: `afterbegin`,
+  BEFOREEND: `beforeend`,
+  AFTEREND: `afterend`
+};
+
+export {POINT_COUNT, PlaceTemplate};
+
+
