+++ src/main.js
@@ -1,3 +1,5 @@
+import {POINT_COUNT, PlaceTemplate} from "./const.js";
+
 import {createTripInfoTemplate} from './view/trip-info.js';
 import {createTripInfoMainTemplate} from './view/trip-info-main.js';
 import {createTripInfoCostTemplate} from './view/trip-info-cost.js';
@@ -6,25 +8,16 @@ import {createFilterTemplate} from './view/filter.js';
 import {createSortTemplate} from './view/sort.js';
 import {createTripDaysTemplate} from './view/trip-days.js';
 import {createTripDayTemplate} from './view/trip-day.js';
-import {createEditEventTemplate} from './view/edit-event.js';
-import {createEventTemplate} from './view/event.js';
-
-const PlaceTemplate = {
-  BEFOREBEGIN: `beforebegin`,
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-  AFTEREND: `afterend`
-};
-
-const EVENT_COUNT = 3;
+import {createPointEditTemplate} from './view/point-edit.js';
+import {createPointTemplate} from './view/point.js';
 
 const tripMainElement = document.querySelector(`.trip-main`);
-const tripEventsElement = document.querySelector(`.trip-events`);
+const tripPointsElement = document.querySelector(`.trip-events`);
 
 const switchMenuElement = tripMainElement.querySelector(`.trip-controls > h2:first-child`);
 const filterElement = tripMainElement.querySelector(`.trip-controls > h2:last-child`);
 
-const softElement = tripEventsElement.querySelector(`:scope > h2:first-child`);
+const softElement = tripPointsElement.querySelector(`:scope > h2:first-child`);
 
 const render = (container, template, place = PlaceTemplate.BEFOREEND) => {
   container.insertAdjacentHTML(place, template);
@@ -40,16 +33,16 @@ render(tripInfoElement, createTripInfoCostTemplate());
 render(switchMenuElement, createSwitchMenuTemplate(), PlaceTemplate.AFTEREND);
 render(filterElement, createFilterTemplate(), PlaceTemplate.AFTEREND);
 render(softElement, createSortTemplate(), PlaceTemplate.AFTEREND);
-render(tripEventsElement, createTripDaysTemplate());
+render(tripPointsElement, createTripDaysTemplate());
 
-const tripDaysElement = tripEventsElement.querySelector(`.trip-days`);
+const tripDaysElement = tripPointsElement.querySelector(`.trip-days`);
 
 render(tripDaysElement, createTripDayTemplate());
 
-const tripEventsList = tripDaysElement.querySelector(`.trip-events__list`);
+const tripPointsList = tripDaysElement.querySelector(`.trip-events__list`);
 
-render(tripEventsList, createEditEventTemplate());
+render(tripPointsList, createPointEditTemplate());
 
-for (let i = 1; i < EVENT_COUNT; i++) {
-  render(tripEventsList, createEventTemplate());
+for (let i = 1; i < POINT_COUNT; i++) {
+  render(tripPointsList, createPointTemplate());
 }
