+++ src/view/point-edit.js
@@ -1,4 +1,4 @@
-export const createEditEventTemplate = () => (
+export const createPointEditTemplate = () => (
   `<li class="trip-events__item">
     <form class="event event--edit" action="#" method="post">
       <header class="event__header">
